back-end/isBack/requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
install==1.3.5
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1

